Added RoleCheck (Admins acess Everything)

+++ ESW_Shelter/Controllers/AnimalTypesController.cs
@@ -9,11 +9,11 @@ using ESW_Shelter.Models;
 
 namespace ESW_Shelter.Controllers
 {
-    public class AnimalTypesController : Controller
+    public class AnimalTypesController : SharedController
     {
         private readonly ShelterContext _context;
 
-        public AnimalTypesController(ShelterContext context)
+        public AnimalTypesController(ShelterContext context) : base(context)
         {
             _context = context;
         }
@@ -21,12 +21,20 @@ namespace ESW_Shelter.Controllers
         // GET: AnimalTypes
         public async Task<IActionResult> Index()
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             return View(await _context.AnimalTypes.ToListAsync());
         }
 
         // GET: AnimalTypes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -43,8 +51,12 @@ namespace ESW_Shelter.Controllers
         }
 
         // GET: AnimalTypes/Create
-        public IActionResult Create()
+        public async Task<IActionResult> CreateAsync()
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             return View();
         }
 
@@ -55,6 +67,10 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("AnimalTypeID,Name")] AnimalType animalType)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(animalType);
@@ -67,6 +83,10 @@ namespace ESW_Shelter.Controllers
         // GET: AnimalTypes/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -87,9 +107,13 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("AnimalTypeID,Name")] AnimalType animalType)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id != animalType.AnimalTypeID)
             {
-                return NotFound();
+                return ErrorNotFoundOrSomeOtherError();
             }
 
             if (ModelState.IsValid)
@@ -118,6 +142,10 @@ namespace ESW_Shelter.Controllers
         // GET: AnimalTypes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -138,6 +166,10 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             var animalType = await _context.AnimalTypes.FindAsync(id);
             _context.AnimalTypes.Remove(animalType);
             await _context.SaveChangesAsync();
