+++ ESW_Shelter/Controllers/ProductsController.cs
@@ -10,11 +10,11 @@ using System.Data.Entity;
 
 namespace ESW_Shelter.Controllers
 {
-    public class ProductsController : Controller
+    public class ProductsController : SharedController
     {
         private readonly ShelterContext _context;
 
-        public ProductsController(ShelterContext context)
+        public ProductsController(ShelterContext context) : base(context)
         {
             _context = context;
         }
@@ -22,7 +22,10 @@ namespace ESW_Shelter.Controllers
         // GET: Products
         public async Task<IActionResult> Index(string searchString, string animalType, string productType)
         {
-
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             var query = from product in _context.Products
                         join productsType in _context.ProductTypes on product.ProductTypeFK equals productsType.ProductTypeID
                         join animalsType in _context.AnimalTypes on product.AnimalTypeFK equals animalsType.AnimalTypeID
@@ -83,6 +86,10 @@ namespace ESW_Shelter.Controllers
         // GET: Products/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -101,6 +108,10 @@ namespace ESW_Shelter.Controllers
         // GET: Products/Create
         public IActionResult Create()
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             ViewBag.AnimalsTypes = _context.AnimalTypes.AsParallel();
             ViewBag.ProductType = _context.ProductTypes.AsParallel();
             return View();
@@ -113,6 +124,10 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ProductID,Name,Quantity,AnimalTypeFK,ProductTypeFK")] Product product)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (ModelState.IsValid)
             {
                 product.AnimalTypeFK = Int32.Parse(Request.Form["AnimalTypeFK"].ToString());
@@ -127,6 +142,10 @@ namespace ESW_Shelter.Controllers
         // GET: Products/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             ViewBag.AnimalsTypes = _context.AnimalTypes.AsParallel();
             ViewBag.ProductType = _context.ProductTypes.AsParallel();
 
@@ -150,6 +169,10 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ProductID,Name,Quantity,AnimalTypeFK,ProductTypeFK")] Product product)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id != product.ProductID)
             {
                 return NotFound();
@@ -181,6 +204,10 @@ namespace ESW_Shelter.Controllers
         // GET: Products/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             if (id == null)
             {
                 return NotFound();
@@ -223,6 +250,10 @@ namespace ESW_Shelter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!GetAutorization(4))
+            {
+                return ErrorNotFoundOrSomeOtherError();
+            }
             var product = await _context.Products.FindAsync(id);
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
